Fix some SpotBug warnings in the unit tests.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1831990 13f79535-47bb-0310-9956-ffa450edef68

+++ test/org/apache/catalina/authenticator/TesterDigestAuthenticatorPerformance.java
@@ -105,7 +105,7 @@ public class TesterDigestAuthenticatorPerformance {
         System.out.println("Average time per request (wall): " +
                 wallTime/(threadCount * requestCount));
 
-        Assert.assertEquals(requestCount * threadCount, totalSuccess);
+        Assert.assertEquals(((long)requestCount) * threadCount, totalSuccess);
     }
 
     @Before
